requirements.txt
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.40
SQLAlchemy_Utils==0.38.3
vk_api==11.9.9
